Add previous & back buttons to webview toolbar
The web view controller uses a nib file for the view. This nib has a
web view and a toolbar. The toolbar has previous and back buttons that
interact with the web view's history.

This completes the Chapter 21 Silver challenge.  The challenge description
said to add the toolbar to the web view controller's view hierarchy. Would
this be easier using the navigation controller's optional toolbar?