Add previous & back buttons to webview toolbar -v2
This does the same thing as the branch chapter21/silver, but in a very
different way.  Instead of adding the toolbar to the webview controller's
hierarchy, it places the buttons on the navigation controller's toolbar.
I don't know which method I prefer. This version has fewer changes from
the parent commit, but is less self-contained.